+++ client/src/utils/validateSignUpForms.ts
